+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 1.6.3
 
-* API update [#96](https://github.com/wvwwvwwv/scalable-concurrent-containers/issues/96): add `HashIndex::{modify, modify_async}`, special thanks to [novacrazy](https://github.com/novacrazy).
+* API update 1: [#96](https://github.com/wvwwvwwv/scalable-concurrent-containers/issues/96) - add `HashIndex::{modify, modify_async}`, special thanks to [novacrazy](https://github.com/novacrazy).
+* API update 2: `Hash*::default()` for any `H: BuildHasher + Default`, by [novacrazy](https://github.com/novacrazy).
 
 1.6.2